refactor: unify utils and exports
