+++ packages/zip-stream/src/archivers/archive-entry.ts
