+++ packages/zip-stream/src/archivers/archive-output-stream.ts
