+++ packages/zip-stream/src/archivers/crc32-stream.ts
