+++ packages/zip-stream/src/archivers/general-purpose-bit.ts
