+++ packages/zip-stream/src/archivers/unix-stat.ts
