+++ packages/zip-stream/src/archivers/zip-archive-entry.ts
