+++ packages/zip-stream/src/compress-commons.ts
