+++ packages/zip-stream/tsdown.config.ts
@@ -1,13 +1,7 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: [
-    "src/index.ts",
-    "src/utils.ts",
-    "src/compress-commons/index.ts",
-    "src/compress-commons/constants.ts",
-    "src/compress-commons/utils.ts",
-  ],
+  entry: ["src/index.ts", "src/utils.ts", "src/constants.ts"],
   dts: true,
   exports: true,
 });
